Prevent failed recursive var lookups from failing templated logic
When using a logical expression for a template variable (for example:
"{{ 'yes' if foo==1 else 'no' }}"), in the case where foo was not yet
defined the template engine would incorrectly evaluate the statement.
This patch fixes the issue by modifying the __contains__() function
to render the variable during the lookup while specifying that undefined
variables fail. If an exception is triggered due to an undefined
variable, __contains__() will now return False instead of True and
the logical operation will not be evaluated until it is referenced later.

Fixes #7786