Start of work to add support for PEX to Ansible
This feature would allow Ansible to package 3rd party dependencies as well
as the entire python executable environment using the pex library. The benefit
of this is that the remote system would not need to install dependencies, and
would not even have to have Python installed, greatly reducing remote setup
time.

This is a WIP and is not ready for primetime, as it is quite a bit slower than
AnsiBallZ right now, especially when adding in any 3rd party deps.