AnsiballZ improvements
Now that we don't need to worry about python-2.4 and 2.5, we can make
some improvements to the way AnsiballZ handles modules.

* Change AnsiballZ wrapper to use import to invoke the module

We need the module to think of itself as a script because it could be
coded as:

    main()

or as:

    if __name__ == '__main__':
        main()

Or even as:

    if __name__ == '__main__':
        random_function_name()

A script will invoke all of those.  Prior to this change, we invoked
a second Python interpreter on the module so that it really was
a script.  However, this means that we have to run python twice (once
for the AnsiballZ wrapper and once for the module).  This change makes
the module think that it is a script (because __name__ in the module ==
'__main__') but it's actually being invoked by us importing the module
code.

* Use zipimporter instead of saving the module to disk

zipimporter can load a python module from within a zipfile.  Use it to
import the module so we don't have to extract the script from the zip
file and save it to disk.

Rename the module in the zip file to __main__.py because zipimpotr
doesn't have the facilities for use to load a file with a different name
than the filename has.

Rename the zip file we save so it's associated with the ansible module
(since the script file is now named __main__.py instead of containing
the name of the ansible module)

An alternative to this would be to use exec but I decided not to use
exec after reading mitsuhiko's article on hidden exec performance
issues: lucumr.pocoo.org/2011/2/1/exec-in-python/

* Rename the name of the AnsiBallZ wrapped module

This makes it obvious that the wrapped module isn't the module file that
we distribute.  It's part of trying to mitigate the fact that the module
is now named __main)).py in tracebacks.

* Monkeypatch __main__.__file__ for backwards compat

Modules could be using __file__.  Monkeypatch __main__ to include
a __file__ attribute

* Fixes module arg passing for and sanity tests and unittests